Add files via upload
AZ_Mart is a retailer selling different products to organizations of different sizes. They have implemented an ERP system to record sales transactions. AZ_Mart has reached out to my course provider for providing Analytics services. As a part of this, I was supposed to deliver a Sales Dashboard in MS Power BI.  
Excel files were the database of the ERP System and individual excel worksheets were tables,I was required to design and develop the dashboard as per Mart's requirements.